Actualizacion esquema de  proyecto
