+++ plantik-app/backend/routes/RegistroVentas.js
@@ -1,0 +1,2 @@
+// routes de registro ventas
+
