+++ plantik-app/frontend/.gitignore
