+++ plantik-app/frontend/package.json
