+++ plantik-app/frontend/public/favicon.ico
