+++ plantik-app/frontend/public/manifest.json
