+++ plantik-app/frontend/src/components/auth/login.css
@@ -1,0 +1,22 @@
+.login-container{
+    margin-top: 3%;
+}
+
+.login-form{
+    
+    border-radius: 50% 20% / 10% 40%;
+    box-shadow: 0 5px 8px 0 rgba(0, 0, 0, 0.2), 0 9px 26px 0 rgba(0, 0, 0, 0.19);
+    padding: 10%;
+}
+
+.login-form h2{
+    text-align: center;
+    color: #333;
+}
+
+.btnSubmit{
+    width: 100%;
+    margin-top: 3%;
+    border-radius: 50% 20% / 10% 40%  !important;
+}
+
