+++ plantik-app/frontend/src/components/gestionVentas/EstadoDeVnta.html
