+++ plantik-app/frontend/src/components/gestionVentas/Plantik.jpeg
