+++ plantik-app/frontend/src/components/maestroProductos/MasterProducts.js
