--- plantik-app/src/components/registroProductos/ProductRegistration.css
@@ -1,8 +1,0 @@
-.product-container{
-    margin-top: 3%;
-}
-
-.regVentas-form h2{
-    text-align: center;
-    color: #333;
-}
